requirements.txt
fastapi 
uvicorn
RPi.GPIO